refactor: organize new step methods in world file
